+++ contrib/experimental/great_expectations_experimental/expectations/expect_column_values_to_be_string_integers_increasing.py
@@ -9,12 +9,135 @@ from great_expectations.core.expectation_validation_result import (
     ExpectationValidationResult,
 )
 from great_expectations.exceptions.exceptions import InvalidExpectationKwargsError
-from great_expectations.execution_engine.execution_engine import ExecutionEngine
+from great_expectations.execution_engine import (
+    ExecutionEngine,
+    PandasExecutionEngine,
+    SparkDFExecutionEngine,
+)
+from great_expectations.execution_engine.execution_engine import (
+    MetricDomainTypes,
+    MetricPartialFunctionTypes,
+)
 from great_expectations.expectations.expectation import ColumnExpectation
+from great_expectations.expectations.metrics import (
+    ColumnMapMetricProvider,
+    column_function_partial,
+    metric_partial,
+)
+from great_expectations.expectations.metrics.import_manager import F, Window, sparktypes
 from great_expectations.expectations.registry import get_metric_kwargs
 from great_expectations.validator.metric_configuration import MetricConfiguration
 
 logger = logging.getLogger(__name__)
+
+
+class ColumnValuesStringIntegersIncreasing(ColumnMapMetricProvider):
+    function_metric_name = "column_values.string_integers.increasing"
+    function_value_keys = ("column", "strictly")
+
+    @column_function_partial(engine=PandasExecutionEngine)
+    def _pandas(self, _column, **kwargs):
+        if all(_column.str.isdigit()) is True:
+            temp_column = _column.astype(int)
+        else:
+            raise TypeError(
+                "Column must be a string-type capable of being cast to int."
+            )
+
+        series_diff = np.diff(temp_column)
+
+        strictly: Optional[bool] = kwargs.get("strictly") or False
+
+        if strictly:
+            return series_diff > 0
+        else:
+            return series_diff >= 0
+
+    @metric_partial(
+        engine=SparkDFExecutionEngine,
+        partial_fn_type=MetricPartialFunctionTypes.WINDOW_FN,
+        domain_type=MetricDomainTypes.COLUMN,
+    )
+    def _spark(
+        cls,
+        execution_engine,
+        metric_domain_kwargs,
+        metric_value_kwargs,
+        metrics,
+        runtime_configuration,
+    ):
+        column_name = metric_domain_kwargs["column"]
+        table_columns = metrics["table.column_types"]
+        column_metadata = [col for col in table_columns if col["name"] == column_name][
+            0
+        ]
+
+        if isinstance(column_metadata["type"], (sparktypes.StringType)):
+            column = F.col(column_name).cast(sparktypes.IntegerType())
+        else:
+            raise TypeError(
+                "Column must be a string-type capable of being cast to int."
+            )
+
+        compute_domain_kwargs = metric_domain_kwargs
+
+        (
+            df,
+            compute_domain_kwargs,
+            accessor_domain_kwargs,
+        ) = execution_engine.get_compute_domain(
+            compute_domain_kwargs, domain_type=MetricDomainTypes.COLUMN
+        )
+
+        if any(np.array(df.select(column.isNull()).collect())):
+            raise TypeError(
+                "Column must be a string-type capable of being cast to int."
+            )
+
+        diff = column - F.lag(column).over(Window.orderBy(F.lit("constant")))
+        diff = F.when(diff.isNull(), 1).otherwise(diff)
+
+        if metric_value_kwargs["strictly"] is True:
+            diff = F.when(diff <= 0, F.lit(False)).otherwise(F.lit(True))
+        else:
+            diff = F.when(diff < 0, F.lit(False)).otherwise(F.lit(True))
+
+        return (
+            np.array(df.select(diff).collect()).reshape(-1)[1:],
+            compute_domain_kwargs,
+            accessor_domain_kwargs,
+        )
+
+    @classmethod
+    def _get_evaluation_dependencies(
+        cls,
+        metric: MetricConfiguration,
+        configuration: Optional[ExpectationConfiguration] = None,
+        execution_engine: Optional[ExecutionEngine] = None,
+        runtime_configuration: Optional[dict] = None,
+    ):
+        """Returns a dictionary of given metric names and their corresponding configuration, specifying the metric
+        types and their respective domains"""
+        dependencies: dict = super()._get_evaluation_dependencies(
+            metric=metric,
+            configuration=configuration,
+            execution_engine=execution_engine,
+            runtime_configuration=runtime_configuration,
+        )
+
+        table_domain_kwargs: dict = {
+            k: v for k, v in metric.metric_domain_kwargs.items() if k != "column"
+        }
+        dependencies["table.column_types"] = MetricConfiguration(
+            metric_name="table.column_types",
+            metric_domain_kwargs=table_domain_kwargs,
+            metric_value_kwargs={
+                "include_nested": True,
+            },
+            metric_dependencies=None,
+        )
+
+        return dependencies
 
 
 class ExpectColumnValuesToBeStringIntegersIncreasing(ColumnExpectation):
@@ -97,10 +220,10 @@ class ExpectColumnValuesToBeStringIntegersIncreasing(ColumnExpectation):
             if not rule(param_value):
                 raise InvalidExpectationKwargsError(error_message)
 
-    def validate_configuration(
-        self, configuration: Optional[ExpectationConfiguration]
-    ) -> bool:
-        return super().validate_configuration(configuration=configuration)
+    # def validate_configuration(
+    #     self, configuration: Optional[ExpectationConfiguration]
+    # ) -> bool:
+    #     return super().validate_configuration(configuration=configuration)
 
     def get_validation_dependencies(
         self,
@@ -114,12 +237,12 @@ class ExpectColumnValuesToBeStringIntegersIncreasing(ColumnExpectation):
             execution_engine=execution_engine,
             runtime_configuration=runtime_configuration,
         )
-
         metric_kwargs = get_metric_kwargs(
             metric_name="column_values.string_integers.increasing.map",
             configuration=configuration,
             runtime_configuration=runtime_configuration,
         )
+
         dependencies["metrics"][
             "column_values.string_integers.increasing"
         ] = MetricConfiguration(
@@ -138,10 +261,72 @@ class ExpectColumnValuesToBeStringIntegersIncreasing(ColumnExpectation):
         execution_engine: ExecutionEngine = None,
     ) -> Dict:
 
-        SIMI = metrics.get("column_values.string_integers.increasing")
-        success = all(SIMI[0])
+        string_integers_increasing = metrics.get(
+            "column_values.string_integers.increasing"
+        )
+
+        success = all(string_integers_increasing[0])
 
         return ExpectationValidationResult(
-            result={"observed_value": np.unique(SIMI[0], return_counts=True)},
+            result={
+                "observed_value": np.unique(
+                    string_integers_increasing[0], return_counts=True
+                )
+            },
             success=success,
         )
+
+    examples = [
+        {
+            "data": {
+                "a": ["0", "1", "2", "3", "3", "9", "11"],
+                "b": ["0", "1", "2", "3", "4", "9", "11"],
+                "c": ["1", "2", "3", "3", "4", "2021-05-01", "test"],
+                "d": ["1", "2", "3", "3", "0", "6", "9"],
+            },
+            "tests": [
+                {
+                    "title": "positive_test_monotonic",
+                    "exact_match_out": False,
+                    "include_in_gallery": True,
+                    "in": {"column": "a", "strictly": False},
+                    "out": {"success": True},
+                },
+                {
+                    "title": "positive_test_strictly",
+                    "exact_match_out": False,
+                    "include_in_gallery": True,
+                    "in": {"column": "b", "strictly": True},
+                    "out": {"success": True},
+                },
+                {
+                    "title": "negative_test_monotonic",
+                    "exact_match_out": False,
+                    "include_in_gallery": True,
+                    "in": {"column": "d", "strictly": False},
+                    "out": {"success": False},
+                },
+                {
+                    "title": "negative_test_strictly",
+                    "exact_match_out": False,
+                    "include_in_gallery": True,
+                    "in": {"column": "a", "strictly": True},
+                    "out": {"success": False},
+                },
+                {
+                    "title": "negative_test_int_cast",
+                    "exact_match_out": False,
+                    "include_in_gallery": True,
+                    "in": {"column": "c", "strictly": False},
+                    "out": {"success": False},
+                },
+            ],
+        }
+    ]
+
+
+if __name__ == "__main__":
+    self_check_report = (
+        ExpectColumnValuesToBeStringIntegersIncreasing().run_diagnostics()
+    )
+    print(json.dumps(self_check_report, indent=2))
